Get highlight working
